refactoring, part2: file rename
